+++ app/src/main/java/com/example/studyapp/ui/components/FullScreenDialog.kt
@@ -19,7 +19,7 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.example.studyapp.R
-import com.example.studyapp.ui.components.study.SubtopicFullScreenDialog
+import com.example.studyapp.ui.screens.subtopics.SubtopicFullScreenDialog
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
